+++ vite.config.ts
@@ -1,7 +1,8 @@
 import { sveltekit } from "@sveltejs/kit/vite";
 import UnoCSS from "unocss/vite";
+import devtoolsJson from "vite-plugin-devtools-json";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-	plugins: [UnoCSS(), sveltekit()],
+	plugins: [devtoolsJson(), UnoCSS(), sveltekit()],
 });